Updated how the rate change value is calculated and displayed, so that the label shows decimal points.

+++ Nutrivurv/Nutrivurv/Weight.swift
@@ -18,7 +18,7 @@ class Weight: ObservableObject {
         }
     }
     
-    @Published var rateChange: Int = 0 {
+    @Published var rateChange: Double = 0 {
         willSet {
             objectWillChange.send(newValue)
         }
